+++ R/plotting.R
@@ -15,7 +15,10 @@
 #' @param event_size An optional numeric vector that adjusts the size of fire event symbols on the plot. Default is \code{c("Scar" = 4, "Injury" = 2, "Pith/Bark" = 1.5)}.
 #' @param rugbuffer_size An optional integer. If the user plots a rug, thiscontrols the amount of buffer whitespace along the y-axis between the rug and the main plot. Must be >= 2.
 #' @param rugdivide_pos Optional integer if plotting a rug. Adjust the placement of the rug divider along the y-axis. Default is 2.
+#'
 #' @return A ggplot object for plotting or manipulation.
+#'
+#' @export
 get_ggplot <- function(x, color_group, color_id, facet_group, facet_id, facet_type="grid", ylabels=TRUE,
                        yearlims=FALSE, composite_rug=FALSE, filter_prop=0.25,
                        filter_min=2, plot_legend=FALSE, event_size=c("Scar" = 4, "Injury" = 2, "Pith/Bark" = 1.5), 
@@ -148,6 +151,8 @@ get_ggplot <- function(x, color_group, color_id, facet_group, facet_id, facet_ty
 #' Plot an fhx object.
 #'
 #' @param ... Arguments passed on to \code{get_ggplot}.
+#'
+#' @export
 plot.fhx <- function(...) {
   print(get_ggplot(...))
 }